+++ packages/cloud/src/metrics/utils/getEventsInRange.ts
@@ -4,4 +4,8 @@ export const getEventsInRange = (
   events: TrackableEvent[],
   fromDate: Date,
   toDate: Date,
-) => events.filter(({ date }) => date >= fromDate && date < toDate)
+) =>
+  events.filter(({ time }) => {
+    const date = time.toDate()
+    return date >= fromDate && date < toDate
+  })